Finalised database convention plumbing, added representative database convention test, just about to add a dogfood test for database conventions to enforce having a matching embedded resource for the database convention. Needs test for db convention itself in addition to nommy dogfood.
