+++ Conventional.Tests/Conventional/Conventions/Database/DogFoodConventions.cs
@@ -1,0 +1,7 @@
+﻿namespace Conventional.Tests.Conventional.Conventions.Database
+{
+    public class DogFoodConventions
+    {
+         
+    }
+}
