+++ Conventional/Conventional.csproj
@@ -67,9 +67,12 @@
     <Compile Include="Conventions\Cecil\MustNotUseDateTimeOffsetNowConventionSpecification.cs" />
     <Compile Include="Conventions\CollectionPropertiesMustBeImmutableConventionSpecification.cs" />
     <Compile Include="Conventions\ConventionSpecification.cs" />
+    <Compile Include="Conventions\Database\AllIdentityColumnsMustBeNamedTableNameId.cs" />
+    <Compile Include="Conventions\Database\DatabaseConventionSpecification.cs" />
     <Compile Include="Conventions\MustHaveANonPublicDefaultConstructorConventionSpecification.cs" />
     <Compile Include="Conventions\MustHaveAppropriateConstructorsConventionSpecification.cs" />
     <Compile Include="Conventions\EnumerablePropertiesMustBeEagerLoadedConventionSpecification.cs" />
+    <Compile Include="Conventions\MustHaveMatchingEmbeddedResourcesConventionSpecification.cs" />
     <Compile Include="Conventions\NotConventionSpecification.cs" />
     <Compile Include="Conventions\OrConventionSpecification.cs" />
     <Compile Include="Conventions\PropertiesMustBeImmutableConventionSpecification.cs" />
@@ -87,6 +90,8 @@
     <Compile Include="Conventions\RequiresACorrespondingImplementationOfConventionSpecification.cs" />
     <Compile Include="Conventions\Solution\MustOnlyContainInformativeCommentsConventionSpecification.cs" />
     <Compile Include="Conventions\Solution\SolutionConventionSpecification.cs" />
+    <Compile Include="DatabaseSpecimen.cs" />
+    <Compile Include="IDatabaseConventionSpecification.cs" />
     <Compile Include="ISolutionConventionSpecification.cs" />
     <Compile Include="StringConstants.cs" />
     <Compile Include="ConventionResult.cs" />
@@ -94,6 +99,7 @@
     <Compile Include="IConventionSpecification.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StringExtensions.cs" />
+    <Compile Include="TheDatabase.cs" />
     <Compile Include="ThisSolution.cs" />
     <Compile Include="WrappedConventionResult.cs" />
   </ItemGroup>
@@ -102,6 +108,9 @@
       <SubType>Designer</SubType>
     </None>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Conventions\Database\AllIdentityColumnsMustBeNamedTableNameId.sql" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
