+++ Conventional/Conventions/Database/AllIdentityColumnsMustBeNamedTableNameId.cs
@@ -1,0 +1,10 @@
+﻿namespace Conventional.Conventions.Database
+{
+    public class AllIdentityColumnsMustBeNamedTableNameIdConventionSpecification : DatabaseConventionSpecification
+    {
+        protected override string FailureMessage
+        {
+            get { return "All identity columns must be named {TableName}Id."; }
+        }
+    }
+}
