+++ Conventional/IDatabaseConventionSpecification.cs
@@ -1,0 +1,7 @@
+namespace Conventional
+{
+    public interface IDatabaseConventionSpecification
+    {
+        ConventionResult IsSatisfiedBy(DatabaseSpecimen databaseSpecimen);
+    }
+}
