+++ Conventional/TheDatabase.cs
@@ -1,0 +1,15 @@
+namespace Conventional
+{
+    public static class TheDatabase
+    {
+        public static DatabaseSpecimen WithConnectionString(string connectionString)
+        {
+            return new DatabaseSpecimen(connectionString);
+        }
+
+        public static ConventionResult MustConformTo(this DatabaseSpecimen databaseSpecimen, IDatabaseConventionSpecification databaseConventionSpecification)
+        {
+            return databaseConventionSpecification.IsSatisfiedBy(databaseSpecimen);
+        }
+    }
+}
